requirements.txt
# Core data processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0

# Text processing
nltk==3.8.1
spacy==3.7.2

# Data validation
pydantic==2.5.2
jsonschema==4.20.0

# File handling
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Cloud storage
boto3==1.34.10
google-cloud-storage==2.14.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1